refactor: remove unused config fields and change TTS rate to float
Removed 3 unused voice_speech config fields:
- voice_speech_chunk_size (duplicate of audio_chunk_size in SystemConfig)
- voice_speech_error_log_threshold (never implemented)
- voice_speech_phrase_time_limit (never implemented)

Changed tts_default_rate from int to float:
- Updated TTSConfig schema: int → float (default 150.0)
- Updated all TTS engine signatures: rate: Optional[int] → Optional[float]
- Updated base TTSEngine abstract method signature
- Enables precise speech rate control (e.g., 175.5 WPM instead of just 175)
- All three engines (pyttsx3, edge, piper) already support float rates

Added docs/not_implemented/timepicker.md:
- Design document for duration picker (hours + minutes)
- Design document for time picker (HH:MM with timezone)
- Future use: voice_time_level fields, weekly_recap_time, scheduled tasks
- Reusable UI component pattern for web config interface

Version bump: 1.1.0 → 1.1.1 (patch release for config cleanup)

+++ bot/core/tts_engines/edge_engine.py
@@ -46,7 +46,7 @@ class EdgeEngine(TTSEngine):
         self,
         text: str,
         voice: Optional[str] = None,
-        rate: Optional[int] = None,
+        rate: Optional[float] = None,
         volume: float = 1.0,
         guild_id: Optional[int] = None,
         **kwargs
