+++ bot/core/tts_engines/piper_engine.py
@@ -57,7 +57,7 @@ class PiperEngine(TTSEngine):
         self,
         text: str,
         voice: Optional[str] = None,
-        rate: Optional[int] = None,
+        rate: Optional[float] = None,
         volume: float = 1.0,
         guild_id: Optional[int] = None,
         **kwargs
